Added more track scores

+++ src/api/UseCases/BusinessObjects/TrackAudioFeatures.cs
@@ -2,15 +2,15 @@
 {
     public class TrackAudioFeatures
     {
-        public double Acousticness { get; set; }
-        public double Danceability { get; set; }
-        public double Energy { get; set; }
-        public double Instrumentalness { get; set; }
-        public double Liveness { get; set; }
-        public double Loudness { get; set; }
-        public double Mode { get; set; }
-        public double Speechiness { get; set; }
-        public double Tempo { get; set; }
-        public double Valence { get; set; }
+        public decimal Acousticness { get; set; }
+        public decimal Danceability { get; set; }
+        public decimal Energy { get; set; }
+        public decimal Instrumentalness { get; set; }
+        public decimal Liveness { get; set; }
+        public decimal Loudness { get; set; }
+        public decimal Mode { get; set; }
+        public decimal Speechiness { get; set; }
+        public decimal Tempo { get; set; }
+        public decimal Valence { get; set; }
     }
 }
